Add sound exposure level calcs
