+++ mhkit/acoustics/__init__.py
@@ -16,4 +16,17 @@ pressure levels, and time or band aggregate spectral data.
 """
 
 from mhkit.acoustics import io, graphics
-from .analysis import *
+from .analysis import (
+    minimum_frequency,
+    sound_pressure_spectral_density,
+    apply_calibration,
+    sound_pressure_spectral_density_level,
+    band_aggregate,
+    time_aggregate,
+)
+from .spl import (
+    sound_pressure_level,
+    third_octave_sound_pressure_level,
+    decidecade_sound_pressure_level,
+)
+from .sel import sound_exposure_level
